Actualizacion de ms-mantenimiento para implementacion de swagger, hateoas, docker, etc, todos los ms del proyecto funcional y testeados

+++ ms-mantenimiento/src/main/java/com/example/ms_mantenimiento/service/MantenimientoService.java
@@ -20,10 +20,9 @@ public class MantenimientoService {
 
     private final MantenimientoRepository repository;
 
-
     public MantenimientoResponse add(MantenimientoRequest request){
 
-        log.info("Agregar informaciond de mantenimiento", keyValue("nombre de la empresa", request.getEmpresa()));
+        log.info("Crear Mantenimiento", keyValue("empresa", request.getEmpresa()));
 
         Mantenimiento mantenimiento1 = new Mantenimiento();
 
@@ -34,42 +33,63 @@ public class MantenimientoService {
 
         Mantenimiento guardado = repository.save(mantenimiento1);
 
-        return mapToResponse(guardado);
+        log.info("Mantenimiento creado correctamente",
+                keyValue("id", guardado.getId())
+        );
 
+        return mapToResponse(guardado);
     }
 
     public MantenimientoResponse findById(Long id){
-        log.info("Obtener Mantenimiento", keyValue("id", id));
+        log.info("Obtener mantenimiento", keyValue("id", id));
         Mantenimiento mantenimiento1 = repository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Mantenimiento no encontrado"));
+        log.info("Mantenimiento encontrado",
+                keyValue("id", mantenimiento1.getId()),
+                keyValue("empresa", mantenimiento1.getEmpresa())
+        );
         return mapToResponse(mantenimiento1);
     }
 
     public List<MantenimientoResponse> getAll(){
+        log.info("Listando mantenimientos");
         return repository.findAll()
                 .stream()
                 .map(m -> mapToResponse(m))
                 .toList();
-
     }
-    public MantenimientoResponse update(Long id , MantenimientoRequest request){
+
+    public MantenimientoResponse update(Long id, MantenimientoRequest request){
+
         Mantenimiento mantenimiento1 = repository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Entrenador no Encontrado"));;
-        log.info("Actualizar Entrenador", keyValue("id", id));
+                .orElseThrow(() -> new EntityNotFoundException("Mantenimiento no encontrado"));
+
+        log.info("Actualizar Mantenimiento", keyValue("id", id));
+
         mantenimiento1.setEmpresa(request.getEmpresa());
         mantenimiento1.setDescripcionMantenimiento(request.getDescripcionMantenimiento());
         mantenimiento1.setPrecio(request.getPrecio());
         mantenimiento1.setFechaMantenimiento(request.getFechaMantenimiento());
-        repository.save(mantenimiento1);
-        return mapToResponse(mantenimiento1);
+
+        Mantenimiento actualizado = repository.save(mantenimiento1);
+
+        return mapToResponse(actualizado);
     }
 
     public void delete(Long id){
-        log.info("Eliminar Entrenador", keyValue("id", id));
+        if (!repository.existsById(id)) {
+            throw new EntityNotFoundException("No se puede eliminar mantenimiento no encontrado");
+        }
+        log.info("Eliminar Mantenimiento", keyValue("id", id));
         repository.deleteById(id);
+        log.info("Mantenimiento eliminado correctamente",
+                keyValue("id", id));
     }
 
     private MantenimientoResponse mapToResponse(Mantenimiento m) {
+        log.info("Mapeando mantenimiento",
+                keyValue("id", m.getId())
+        );
         return MantenimientoResponse.builder()
                 .id(m.getId())
                 .empresa(m.getEmpresa())
